+++ packages/meeting_place_chat/test/utils/repository/channel_repository_impl.dart
@@ -54,19 +54,6 @@ class ChannelRepositoryImpl implements ChannelRepository {
   }
 
   @override
-  Future<Channel?> findChannelByOfferLink(String offerLink) async {
-    final permanentChanneDid = await _storage.get<String>(
-      '$channelOfferLinkPrefix$offerLink',
-    );
-    if (permanentChanneDid == null) return null;
-
-    final channel = await _storage.get<String>(
-      '$channelPrefix$permanentChanneDid',
-    );
-    return Channel.fromJson(json.decode(channel!) as Map<String, dynamic>);
-  }
-
-  @override
   Future<void> updateChannel(Channel channel) async {
     return _writeChannel(channel);
   }