+++ packages/meeting_place_control_plane/lib/src/command/query_offer/query_offer_output.dart
@@ -43,3 +43,7 @@ class SuccessQueryOfferCommandOutput extends QueryOfferCommandOutput {
 }
 
 class NullQueryOfferCommandOutput extends QueryOfferCommandOutput {}
+
+class LimitExceededQueryOfferCommandOutput extends QueryOfferCommandOutput {}
+
+class ExpiredQueryOfferCommandOutput extends QueryOfferCommandOutput {}