+++ packages/meeting_place_chat/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.0.1-dev.10
+
+ - **FIX**: apply configured retry count and return network_error code in case of connection error (#16).
+
 ## 0.0.1-dev.9
 
  - Update a dependency to the latest release.